re-structuring attendance files
