+++ src/components/Attendance/CheckboxCols.js
@@ -1,12 +1,11 @@
 import React, { useContext, useReducer, useEffect } from 'react';
 import RosterContext from '../../context/roster-context';
-import rosterReducer, { TOGGLE_ATTENDANCE } from './../../reducers/roster';
-import groupsReducer, { GROUP_CHECKBOXES } from './../../reducers/groups';
+import rosterReducer, { TOGGLE_ATTENDANCE } from '../../reducers/roster';
+import groupsReducer, { GROUP_CHECKBOXES } from '../../reducers/groups';
 import Form from 'react-bootstrap/Form';
-import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 
-const AttendanceSelector = () => {
+const CheckboxCols = () => {
 
     const { rosterState } = useContext(RosterContext);
     const rosterDispatch = useReducer(rosterReducer, [])[1];
@@ -27,8 +26,8 @@ const AttendanceSelector = () => {
         })
     };
 
-    const CheckboxCols = ({ groups }) => (
-        groups.map(({ group, orderClasses }, i) => (
+    return (
+        groupedCheckboxRoster.map(({ group, orderClasses }, i) => (
             <Col className={`col-12 col-sm-6 col-md-4 ${orderClasses}`} key={`form-col-${i}`}>
                 <Form>
                     { group.map((student) => (
@@ -45,26 +44,6 @@ const AttendanceSelector = () => {
             </Col>
         ))
     );
-
-    const CheckboxList = () => {
-        return (
-            rosterState.roster ?
-            <React.Fragment>
-                <h3>Mark Absent Students</h3>
-                <Row>
-                    <CheckboxCols groups={groupedCheckboxRoster} />
-                </Row>
-            </React.Fragment>
-            :
-            <React.Fragment />
-        )
-    };
-
-    return (
-        <React.Fragment>
-            <CheckboxList />
-        </React.Fragment>
-    );
 };
 
-export default AttendanceSelector;
+export default CheckboxCols;
